requirements.txt
rich~=13.5
python-dotenv~=1.0
requests~=2.31
